requirements.txt
openai
python-dotenv
gradio
wandb
gradio_multimodalchatbot
llama-index